# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Project Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProjectStatus = typeof ProjectStatus[keyof typeof ProjectStatus];


export const ProjectStatus = {
  active: 'active',
  archived: 'archived',
} as const;

export interface Project {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  status: ProjectStatus;
  color: string;
  createdAt: string;
  updatedAt: string;
}

export type ProjectInputStatus = typeof ProjectInputStatus[keyof typeof ProjectInputStatus];


export const ProjectInputStatus = {
  active: 'active',
  archived: 'archived',
} as const;

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  status?: ProjectInputStatus;
  color?: string;
}

export type ProjectUpdateStatus = typeof ProjectUpdateStatus[keyof typeof ProjectUpdateStatus];


export const ProjectUpdateStatus = {
  active: 'active',
  archived: 'archived',
} as const;

export interface ProjectUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  description?: string | null;
  status?: ProjectUpdateStatus;
  color?: string;
}

export interface ProjectProgress {
  projectId: number;
  total: number;
  todo: number;
  inProgress: number;
  done: number;
  completionPercent: number;
  overdue: number;
}

export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];


export const TaskStatus = {
  todo: 'todo',
  in_progress: 'in_progress',
  done: 'done',
} as const;

export type TaskPriority = typeof TaskPriority[keyof typeof TaskPriority];


export const TaskPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export type MemberRole = typeof MemberRole[keyof typeof MemberRole];


export const MemberRole = {
  admin: 'admin',
  member: 'member',
} as const;

export interface Member {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  avatarUrl?: string | null;
  role: MemberRole;
  createdAt: string;
}

export interface Task {
  id: number;
  projectId: number;
  title: string;
  /** @nullable */
  description?: string | null;
  status: TaskStatus;
  priority: TaskPriority;
  /** @nullable */
  assigneeId?: number | null;
  assignee?: Member;
  project?: Project;
  /** @nullable */
  dueDate?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type TaskInputStatus = typeof TaskInputStatus[keyof typeof TaskInputStatus];


export const TaskInputStatus = {
  todo: 'todo',
  in_progress: 'in_progress',
  done: 'done',
} as const;

export type TaskInputPriority = typeof TaskInputPriority[keyof typeof TaskInputPriority];


export const TaskInputPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface TaskInput {
  projectId: number;
  /** @minLength 1 */
  title: string;
  description?: string;
  status?: TaskInputStatus;
  priority?: TaskInputPriority;
  /** @nullable */
  assigneeId?: number | null;
  /** @nullable */
  dueDate?: string | null;
}

export type TaskUpdateStatus = typeof TaskUpdateStatus[keyof typeof TaskUpdateStatus];


export const TaskUpdateStatus = {
  todo: 'todo',
  in_progress: 'in_progress',
  done: 'done',
} as const;

export type TaskUpdatePriority = typeof TaskUpdatePriority[keyof typeof TaskUpdatePriority];


export const TaskUpdatePriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface TaskUpdate {
  /** @minLength 1 */
  title?: string;
  /** @nullable */
  description?: string | null;
  status?: TaskUpdateStatus;
  priority?: TaskUpdatePriority;
  /** @nullable */
  assigneeId?: number | null;
  /** @nullable */
  dueDate?: string | null;
}

export type MemberInputRole = typeof MemberInputRole[keyof typeof MemberInputRole];


export const MemberInputRole = {
  admin: 'admin',
  member: 'member',
} as const;

export interface MemberInput {
  /** @minLength 1 */
  name: string;
  email: string;
  avatarUrl?: string;
  role?: MemberInputRole;
}

export type MemberUpdateRole = typeof MemberUpdateRole[keyof typeof MemberUpdateRole];


export const MemberUpdateRole = {
  admin: 'admin',
  member: 'member',
} as const;

export interface MemberUpdate {
  /** @minLength 1 */
  name?: string;
  email?: string;
  /** @nullable */
  avatarUrl?: string | null;
  role?: MemberUpdateRole;
}

export interface MemberWorkload {
  memberId: number;
  total: number;
  todo: number;
  inProgress: number;
  done: number;
  overdue: number;
}

export interface DashboardSummary {
  totalProjects: number;
  activeProjects: number;
  totalTasks: number;
  completedTasks: number;
  overdueTasks: number;
  inProgressTasks: number;
  completionRate: number;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  task_created: 'task_created',
  task_updated: 'task_updated',
  task_completed: 'task_completed',
  project_created: 'project_created',
  project_updated: 'project_updated',
} as const;

export type ActivityItemEntityType = typeof ActivityItemEntityType[keyof typeof ActivityItemEntityType];


export const ActivityItemEntityType = {
  task: 'task',
  project: 'project',
} as const;

export interface ActivityItem {
  id: string;
  type: ActivityItemType;
  title: string;
  entityType: ActivityItemEntityType;
  entityId: number;
  /** @nullable */
  projectName?: string | null;
  /** @nullable */
  assigneeName?: string | null;
  createdAt: string;
}

export type ListProjectsParams = {
status?: ListProjectsStatus;
};

export type ListProjectsStatus = typeof ListProjectsStatus[keyof typeof ListProjectsStatus];


export const ListProjectsStatus = {
  active: 'active',
  archived: 'archived',
} as const;

export type ListTasksParams = {
projectId?: number;
assigneeId?: number;
status?: ListTasksStatus;
priority?: ListTasksPriority;
};

export type ListTasksStatus = typeof ListTasksStatus[keyof typeof ListTasksStatus];


export const ListTasksStatus = {
  todo: 'todo',
  in_progress: 'in_progress',
  done: 'done',
} as const;

export type ListTasksPriority = typeof ListTasksPriority[keyof typeof ListTasksPriority];


export const ListTasksPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export type GetRecentActivityParams = {
limit?: number;
};