# lib/api-zod/src/generated/types/memberUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Project Management API
 * OpenAPI spec version: 0.1.0
 */
import type { MemberUpdateRole } from './memberUpdateRole';

export interface MemberUpdate {
  /** @minLength 1 */
  name?: string;
  email?: string;
  /** @nullable */
  avatarUrl?: string | null;
  role?: MemberUpdateRole;
}